+++ src/number_theory/multiplicative_inverse.cpp
@@ -7,7 +7,7 @@
         gcd(a, md) == 1のときのみ乗法逆元が存在
 */
 
-using i64 = long long int;
+using i64 = int64_t;
 
 // a : 元, md : 法
 i64 multi_inv(i64 a, i64 md) {
